+++ src/autoNumeric.js
@@ -2,7 +2,7 @@
 * autoNumeric.js
 * @author: Bob Knothe
 * @contributors: Sokolov Yura and other Github users
-* @version: 2.0 - 2016-11-16 UTC-10 23:00
+* @version: 2.0 - 2016-11-24 GMT UTC-10 21:15
 *
 * Created by Robert J. Knothe on 2009-08-09. Please report any bugs to https://github.com/BobKnothe/autoNumeric
 *
@@ -1876,15 +1876,12 @@ if (typeof define === 'function' && define.amd) {
                         if (value === 'true' || value === 'false') {
                             settings[key] = Boolean(value === 'true');
                         }
-                        if (typeof value === 'number' && key !== 'aScale') {
-                            settings[key] = value.toString();
-                        }
                     });
 
-                    if (settings.aScale !== null) {
-                        settings.scaleFactor = +settings.aScale[0];
-                        settings.scaleDecimal = (settings.aScale[1]) ? +settings.aScale[1] : null;
-                        settings.scaleSuffix = (settings.aScale[2]) ? settings.aScale[2] : '';
+                    if (settings.scaleDivisor !== null) {
+                        settings.scaleDivisor = +settings.scaleDivisor;
+                        settings.scaleDecimal = (settings.scaleDecimal) ? +settings.scaleDecimal : null;
+                        settings.scaleSymbol = (settings.scaleSymbol) ? settings.scaleSymbol : '';
                     }
 
                     // Save our new settings
@@ -1962,7 +1959,7 @@ if (typeof define === 'function' && define.amd) {
                                 if (settings.eDec !== null && settings.aStor) {
                                     settings.rawValue = autoSave($this, settings, 'get');
                                 }
-                                if (settings.aScale && settings.aStor) {
+                                if (settings.scaleDivisor && settings.aStor) {
                                     settings.rawValue = autoSave($this, settings, 'get');
                                 }
                                 if (!settings.aStor) {
@@ -2032,7 +2029,7 @@ if (typeof define === 'function' && define.amd) {
                         if ($settings.eDec) {
                             $settings.mDec = $settings.eDec;
                             $this.autoNumeric('set', $settings.rawValue);
-                        } else if ($settings.aScale) {
+                        } else if ($settings.scaleDivisor) {
                             $settings.mDec = $settings.oDec;
                             $this.autoNumeric('set', $settings.rawValue);
                         } else if ((result = autoStrip($this.val(), $settings)) !== $settings.rawValue) {
@@ -2179,11 +2176,11 @@ if (typeof define === 'function' && define.amd) {
                             if (checkEmpty(value, $settings) === null && minTest && maxTest) {
                                 value = fixNumber(value, $settings.aDec, $settings.aNeg);
                                 $settings.rawValue = value;
-                                if ($settings.aScale) {
-                                    value = value / $settings.scaleFactor;
+                                if ($settings.scaleDivisor) {
+                                    value = value / $settings.scaleDivisor;
                                     value = value.toString();
                                 }
-                                $settings.mDec = ($settings.aScale && $settings.aScale[1]) ? +$settings.scaleDecimal : $settings.mDec;
+                                $settings.mDec = ($settings.scaleDivisor && $settings.scaleDecimal) ? +$settings.scaleDecimal : $settings.mDec;
                                 value = autoRound(value, $settings);
                                 value = presentNumber(value, $settings);
                             } else {
@@ -2208,7 +2205,7 @@ if (typeof define === 'function' && define.amd) {
                             groupedValue = autoGroup(value, $settings);
                         }
                         if (groupedValue !== origValue) {
-                            groupedValue = ($settings.scaleSuffix) ? groupedValue + $settings.scaleSuffix : groupedValue;
+                            groupedValue = ($settings.scaleSymbol) ? groupedValue + $settings.scaleSymbol : groupedValue;
                             $this.val(groupedValue);
                         }
                         if (groupedValue !== holder.inVal) {
@@ -2313,10 +2310,10 @@ if (typeof define === 'function' && define.amd) {
                 const strip = $this.autoNumeric('get');
                 settings = $.extend(settings, options);
 
-                if (settings.aScale !== null) {
-                    settings.scaleFactor = +settings.aScale[0];
-                    settings.scaleDecimal = (settings.aScale[1]) ? +settings.aScale[1] : null;
-                    settings.scaleSuffix = (settings.aScale[2]) ? settings.aScale[2] : '';
+                if (settings.scaleDivisor !== null) {
+                    settings.scaleDivisor = +settings.scaleDivisor;
+                    settings.scaleDecimal = (settings.scaleDecimal) ? +settings.scaleDecimal : null;
+                    settings.scaleSymbol = (settings.scaleSymbol) ? settings.scaleSymbol : '';
                 }
                 settings = originalSettings(settings);
                 getHolder($this, settings, true);
@@ -2363,25 +2360,25 @@ if (typeof define === 'function' && define.amd) {
                 if (value !== '') {
                     const [minTest, maxTest] = autoCheck(value, settings);
                     if (minTest && maxTest) {
-                        if ($input && (!settings.eDec || !settings.aScale)) {
+                        if ($input && (!settings.eDec || !settings.divisor)) {
                             settings.rawValue = value;
                         }
 
                         // checks if the value falls within the min max range
                         if ($input || isInArray($this.prop('tagName').toLowerCase(), settings.tagList)) {
-                            if (settings.aScale && !settings.onOff) {
-                                value = value / settings.scaleFactor;
+                            if (settings.scaleDivisor && !settings.onOff) {
+                                value = value / settings.scaleDivisor;
                                 value = value.toString();
                                 settings.mDec = settings.scaleDecimal;
                             }
                             value = autoRound(value, settings);
-                            if (settings.eDec === null && settings.aScale === null) {
+                            if (settings.eDec === null && settings.scaleDivisor === null) {
                                 settings.rawValue = value;
                             }
                             value = presentNumber(value, settings);
                             value = autoGroup(value, settings);
                         }
-                        if (settings.aStor && (settings.eDec !== null || settings.aScale !== null)) {
+                        if (settings.aStor && (settings.eDec !== null || settings.scaleDivisor !== null)) {
                             autoSave($this, settings, 'set');
                         }
                     } else {
@@ -2402,8 +2399,8 @@ if (typeof define === 'function' && define.amd) {
                     return $this.val('');
                 }
 
-                if (!settings.onOff && settings.scaleSuffix) {
-                    value = value + settings.scaleSuffix;
+                if (!settings.onOff && settings.scaleSymbol) {
+                    value = value + settings.scaleSymbol;
                 }
                 if ($input) {
                     return $this.val(value);
@@ -2472,7 +2469,7 @@ if (typeof define === 'function' && define.amd) {
                 throwError(`The "<${$this.prop('tagName').toLowerCase()}>" tag is not supported by autoNumeric`, settings.debug);
             }
 
-            if (settings.eDec || settings.aScale) {
+            if (settings.eDec || settings.scaleDivisor) {
                 value = settings.rawValue;
             } else {
                 if (!((/\d/).test(value) || Number(value) === 0) && settings.wEmpty === 'focus') {
@@ -2668,16 +2665,30 @@ if (typeof define === 'function' && define.amd) {
          */
         eDec: null,
 
-        /* Scaled number displayed when input does not have focus example with the following:
-         * {aScale: ["1000", "0", "K"]}  => with focus "1,000.00" without focus "1K"
-         * ["divisor", "decimal places", "symbol"]
-         * divisor value - does not need to be whole number - please understand that Javascript has limited accuracy in math
-         * the "get" method returns the full value and scaled value.
-         * decimal places "optional" when not in focus - if omitted the decimal places will be the same when the input has focus
-         * Symbol "optional" displayed when the input does not have focus - NOTE: if a symbol is used you MUST also specify the decimal places
-         * value must be enclosed in quotes example mDec: "3"
+        /* The next three options (scaleDivisor, scaleDecimal & scaleSymbol) handle scaling of the input when the input does not have focus
+         * Please note that the non-scaled value is held in data and it is advised that you use the "aStore" option to ensure retaining the value         * ["divisor", "decimal places", "symbol"]
+         * Example: with the following options set {scaleDivisor: '1000', scaleDecimal: '1', scaleSymbol: ' K'}
+         * Example: focusin value "1,111.11" focusout value "1.1 K"
          */
-        aScale: null,
+
+        /* scaleDivisor devides the on focus value and places the result in the input on focusout
+         * example {scaleDivisor: '1000'} or <input data-scale-divisor="1000">
+         * the divisor value - does not need to be whole number but please understand that Javascript has limited accuracy in math
+         * the "get" method returns the full value.
+         */
+        scaleDivisor: 1000,
+
+        /*
+         * scaledDecimal option is the number of decimal place when not in focus - for this to function scaledDivisor must not be null
+         * this is "optional" if omitted the decimal places will be the same when the input has focus
+         */
+        scaleDecimal: 1,
+
+        /*
+         * scaledSymbol option is a symbol placed as a suffix when not in focus.
+         * this is "optional"
+         */
+        scaleSymbol: 'x',
 
         /* Set to true to allow the eDec value to be saved with sessionStorage
          * if ie 6 or 7 the value will be saved as a session cookie
@@ -2685,10 +2696,10 @@ if (typeof define === 'function' && define.amd) {
         aStor: false,
 
         /* method used for rounding
-         * mRound: "s", Round-Half-Up Symmetric (default)
+         * mRound: "S", Round-Half-Up Symmetric (default)
          * mRound: "A", Round-Half-Up Asymmetric
          * mRound: "s", Round-Half-Down Symmetric (lower case s)
-         * mRound: "A", Round-Half-Down Asymmetric (lower case a)
+         * mRound: "a", Round-Half-Down Asymmetric (lower case a)
          * mRound: "B", Round-Half-Even "Bankers Rounding"
          * mRound: "U", Round Up "Round-Away-From-Zero"
          * mRound: "D", Round Down "Round-Toward-Zero" - same as truncate